refactor: configured profile to use the new hooks
